net_processing: fix potential integer underflow in reverse loop
When headers vector is empty, headers.size() - 1 causes underflow
since size() returns size_t (unsigned). Cast to int to prevent this.

+++ src/net_processing.cpp
@@ -3405,7 +3405,7 @@ void PeerManagerImpl::ProcessGetCFCheckPt(CNode& node, Peer& peer, DataStream& v
 
     // Populate headers.
     const CBlockIndex* block_index = stop_index;
-    for (int i = headers.size() - 1; i >= 0; i--) {
+    for (int i = static_cast<int>(headers.size()) - 1; i >= 0; i--) {
         int height = (i + 1) * CFCHECKPT_INTERVAL;
         block_index = block_index->GetAncestor(height);
